+++ src/youdotcom/runs.py
@@ -52,7 +52,7 @@ class Runs(BaseSDK):
         if server_url is not None:
             base_url = server_url
         else:
-            base_url = models.AGENTS_RUNS_OP_SERVERS[0]
+            base_url = self._get_url(base_url, url_variables)
 
         if not isinstance(request, BaseModel):
             request = utils.unmarshal(request, models.AgentsRunsRequest)
@@ -190,7 +190,7 @@ class Runs(BaseSDK):
         if server_url is not None:
             base_url = server_url
         else:
-            base_url = models.AGENTS_RUNS_OP_SERVERS[0]
+            base_url = self._get_url(base_url, url_variables)
 
         if not isinstance(request, BaseModel):
             request = utils.unmarshal(request, models.AgentsRunsRequest)